#14 refactor: Modify Vue Router Setting
