+++ src/components/user/SignUpForm.vue
@@ -1,0 +1,22 @@
+<template>
+    <div class="container">
+        <p>hi</p>
+    </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+    
+})
+
+</script>
+
+<style>
+
+p {
+    color: white;
+}
+</style>
+
