+++ src/views/HomeView.vue
@@ -1,11 +1,18 @@
 <template>
-  <div></div>
+  <div>
+    <StoreBoard />
+  </div>
 </template>
 
-<script>
-export default {
-  name: 'HomeView'
-}
+<script lang="ts">
+import { defineComponent } from 'vue';
+import StoreBoard from '@/components/store/StoreBoard.vue';
+
+export default defineComponent({
+  components: {
+    StoreBoard
+  }
+})
 </script>
 
 <style scoped>
