+++ src/views/SignUpView.vue
@@ -1,0 +1,22 @@
+<template>
+    <div class="container">
+        <SignUpForm />
+    </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+import SignUpForm from '../components/user/SignUpForm.vue'
+
+export default defineComponent({
+
+    components: {
+        SignUpForm
+    }
+
+})
+
+</script>
+
+<style>
+</style>
